requirements.txt
atomicwrites==1.4.0
attrs==20.3.0
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
colorama==0.4.4
Deprecated==1.2.12
idna==2.10
iniconfig==1.1.1
packaging==20.9
pluggy==0.13.1
py==1.10.0
pycparser==2.20
PyGithub==1.55
PyJWT==2.1.0
PyNaCl==1.4.0
pyparsing==2.4.7
pytest==6.2.4
requests==2.25.1
six==1.15.0
toml==0.10.2
urllib3==1.26.4
wrapt==1.12.1
